wind/requirements.txt
--index-url https://download.pytorch.org/whl/cpu

Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
SQLAlchemy==2.0.21
Alembic>=1.12.0

# AI and ML packages
torch==2.0.1+cpu
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community==0.0.38
google-generativeai==0.2.2

# Celery and Redis
redis==5.0.0
celery==5.3.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
eventlet==0.33.3
python-socketio>=5.9.0
gunicorn>=21.0.0
python-dateutil==2.8.2

# Optional integrations
twilio==8.8.0
PyPDF2==3.0.1
